% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.grey.pal.R
\name{cb.grey.pal}
\alias{cb.grey.pal}
\title{Color platte for color bling ggplot2 plot}
\usage{
cb.grey.pal(...)
}
\value{
a vector of colors for ploting
}
\description{
Color platte for color bling ggplot2 plot
}
