% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_vec.R
\name{compute.min_mean_sd_max}
\alias{compute.min_mean_sd_max}
\title{function for computing mean, DS, max and min values}
\usage{
compute.min_mean_sd_max(x)
}
\arguments{
\item{x}{}
}
\description{
function for computing mean, DS, max and min values
}
