% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summary_vec.R
\name{compute_q025_q25_q50_q75_q975}
\alias{compute_q025_q25_q50_q75_q975}
\title{define the summary function}
\usage{
compute_q025_q25_q50_q75_q975(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
define the summary function
}
