% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_coords.R
\name{conv_ISPTM_LongLat}
\alias{conv_ISPTM_LongLat}
\title{# Fonction de conversion d'un format ddmm.m en degr?s d?cimaux}
\usage{
conv_ISPTM_LongLat(x)
}
\arguments{
\item{x}{}
}
\description{
# Fonction de conversion d'un format ddmm.m en degr?s d?cimaux
}
