% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.std.R
\name{cov.std.grid}
\alias{cov.std.grid}
\title{Standardize the continuous covariates of the grid based on the observation}
\usage{
cov.std.grid(x, mean_x_obs, sd_x_obs)
}
\arguments{
\item{sd_x_obs}{}
}
\description{
Standardize the continuous covariates of the grid based on the observation
}
