% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_std.R
\name{cov_std_grid}
\alias{cov_std_grid}
\title{Standardize the continuous covariates of the grid based on the observation}
\usage{
cov_std_grid(x, mean_x_obs, sd_x_obs)
}
\arguments{
\item{sd_x_obs}{}
}
\description{
Standardize the continuous covariates of the grid based on the observation
}
