% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_plot_legend.R
\name{extract_plot_legend}
\alias{extract_plot_legend}
\title{Extract a legend from a ggplot object
Extract a legend from a ggplot object and save it if Save argument set to TRUE}
\usage{
extract_plot_legend(
  gplot,
  width = 5,
  height = 5,
  device = "pdf",
  Save = FALSE,
  path_figure = NULL
)
}
\arguments{
\item{gplot}{a ggplot2 object}

\item{width}{a numeric value}

\item{height}{a numeric value}

\item{device}{a charchacter taking the form of pdf or png}

\item{Save}{a boolean}

\item{path_figure}{a directory or relative directory as charchacter}
}
\value{
the legend of a ggplot2 object
}
\description{
Extract a legend from a ggplot object
Extract a legend from a ggplot object and save it if Save argument set to TRUE
}
