% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.NA.gaps.R
\name{fill.NA.gaps}
\alias{fill.NA.gaps}
\title{Fill NA values with previous value in a vector}
\usage{
fill.NA.gaps(x, firstBack = FALSE)
}
\arguments{
\item{firstBack}{}
}
\description{
Fill NA values with previous value in a vector
}
