% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NA_gaps.R
\name{fill_NA_gaps}
\alias{fill_NA_gaps}
\title{Fill NA values with previous value in a vector}
\usage{
fill_NA_gaps(x, firstBack = FALSE)
}
\arguments{
\item{firstBack}{}
}
\value{

}
\description{
Fill NA values with previous value in a vector
}
