% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstallPackages.R
\name{InstallPackages}
\alias{InstallPackages}
\title{Install package if not present}
\usage{
InstallPackages(Packages)
}
\arguments{
\item{Packages}{}
}
\value{

}
\description{
Install package if not present
}
