% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_plotter.R
\name{list_plotter}
\alias{list_plotter}
\title{Function to plot palette within a list}
\usage{
list_plotter(color_list, names, package_name)
}
\arguments{
\item{package_name}{}
}
\description{
Function to plot palette within a list
}
