% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MCMC_diag.R
\name{plot_MCMC_diag}
\alias{plot_MCMC_diag}
\title{Function to make mcmc diag plots from a data.frame}
\usage{
plot_MCMC_diag(df_samples, suffix_files = NULL, SavePlots = TRUE,
  ShowPlots = TRUE, density = TRUE, autocorr = TRUE,
  trace_plot = TRUE, Rhat = TRUE, pair_plot = TRUE,
  geweke_plot = FALSE)
}
\description{
Function to make mcmc diag plots from a data.frame
}
