% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.df.R
\name{round_df}
\alias{round_df}
\title{Round numeric values of a data.frame}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{an integer}
}
\value{

}
\description{
Round numeric values of a data.frame
}
