% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ggplot2_theme.R
\name{simple_ggplot2_theme}
\alias{simple_ggplot2_theme}
\title{My simple ggplot2 theme}
\usage{
simple_ggplot2_theme(base_size = 11, base_family = "")
}
\value{

}
\description{
My simple ggplot2 theme
}
