% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wopen.R
\name{wopen}
\alias{wopen}
\title{Open a windows explorer of the relative path
Save plot with specific needs i.e. plot for a Beamer presentation}
\usage{
wopen(relative_path)
}
\arguments{
\item{relative_path}{a charachter string for the directory where to save the plot}
}
\description{
Open a windows explorer of the relative path
Save plot with specific needs i.e. plot for a Beamer presentation
}
