% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.df.R
\name{round.df}
\alias{round.df}
\title{Round numeric values of a data.frame}
\usage{
\method{round}{df}(df, digits)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{an integer}
}
\value{
a rounded data.frame
}
\description{
Round numeric values of a data.frame
}
